Add files via upload
Excel is the go to tool for data analysis and representation after SQL. A versatile data analyst has a couple of these tools in his arsenal to interpret data as best as possible. This coffee shop data takes the keep simple approach, in which I analyze coffee sales over a period, from the size of coffee sold and loyalty membership.